✨ Rewrite toolchains
This commit splits the toolchains into finer grained heteroteneous and
non-heterogeneous toolchains. This will make it easier to add more
toolchains in the future. A next step might be merging
ll_bootstrap_library into ll_library and differentiating only at the
toolchain level. The new bzlmod setup may benefit from this as well, as
we should be able to bootstrap rules_ll more effectively with
modularized dependencies for heterogeneous applications.