💔 Remove local_library_path
Removes os detection and the local_library_path repository. Adding
external dependencies to the build is now done via new LL_CFLAGS and
LL_LDFLAGS environment variables.

Introduces a showcase nix-shell which we'll rework to a nix flake in the
future. Currently not including glibc correctly but everything else is
now fully hermetic.