🤯 Move CUDA to flake
Removes CUDA from the ll toolchain and instead fetches it as part of the
flake. Since CUDA is closed-source it doesn't really matter whether we
import it via Bazel or Nix, but the latter is more maintainable and the
new infrastructure for it makes it easier to support user-defined
versions in the future.