🎯 Use NIX_CFLAGS_COMPILE_FOR_TARGET
Quality of life change. Removes the duplicate includes that
NIX_CFLAGS_COMPILE has and removes the -frandom-seed flag for better
caching behavior when we change the flake without changing dependencies.

If we want to use -frandom-seed we should make it file-dependent and add
it via Bazel so that we get better caching behavior.