⬆️  Update ROCm-CompilerSupport to 641da67 (#44)
This introduces causes a few new warnings to be raised. These went
undetected for a long time and we need to find a way to fix them.
Unfortunately the naive approach of switching to non-deprecated symbols
didn't work, so we'll have to live with these warnings for a while until
fixes are available upstream.