✨ Use custom-built Bazel
The variant with bazelisk didn't work well with nix-built multi-layer
remote execution images. Switch to an overridden Bazel from nixpkgs and
add logic to create such images.

The images about 2.6GB in size which is almost as small as we can get
for an image containing Bazel and a Clang/LLVM C++ toolchain. We might
be able to shave off ~250MB by building python with that same toolchain
instead of building it with GCC.

Signed-off-by: Aaron Siddhartha Mondal <aaron@eomii.org>