🌱 Migrate from Bazelisk to Bazel (#80)
This is an unintuitive change, but it's necessary to simplify the Bazel
wrapping process as preparation for remote execution. We now pin the
Bazel version through nix instead of Bazelisk. This means that the
reproducible Bazel variant from nixpgs is fetched during environment
startup instead of during the first build invocation.

Using different Bazel versions now requires updating Bazel in nixpkgs,
but we planned on doing this anyways.

Removes the now obsolete and incompatible legacy setup from the docs.