🌯 Wrap Bazel (#82)
Create an environment for bazel that is easily importable in container
images. With this we no longer need to explicitly import the C++
toolchain packages as they are now part of the Bazel wrapper. The
current implementation is very inflexible but provides a starting
point for further improvements towards a proper bazel-wrapper.