🛠️ Introduce new ll development tool (#87)
Scripts were starting to pollute the root directory. With the recently
introduced toolchain generation we have so many specialized workflows
that it makes sense to create a dedicated tool to aggregate them. Since
we've previously removed the ll tool as a user-level tool, we can
reintroduce it as a development tool.

1. `ll docs` is the new documentation generator.
2. `ll overlay` is the tool to create the overlay patches.
3. `ll module` is the bzlmod release tool.
4. `ll rbe` is the previous rbegen tool.