Switch base toolchains to LRE
This commit migrates the base toolchains to NativeLink with LRE.

Due to issues with modules the new remote toolchains are't fully
functional, but non-module LLVM parts of it are.

We can now use `ll up` to set up a NativeLink LRE deployment in a local
K8s cluster to debug against a proper remote exec setup.

Changes the nix-bundled bazel to bazelisk. This is easier to get working
with remote execution and is independently updateable.