🔮 Sync llvm and rocm fork
This moves us to an LLVM version that is compatible with comgr and
device-libs from the ROCm LLVM fork. This way we can keep depending
mainly on upstream LLVM but still get to use newer versions of AMD's
libraries.

The version bumps in this commit are tuned and tested to pass the amdgpu
examples. The clr commit lags behind a bit, but the next commit would
cause segfaults due to a bumped default code object version.