🍦 Bring CUDA setup in sync with recent nixpkgs
This is a subtle bug. The CUDA_TOOLKIT changed on the nix side and no
longer contained the actual cuda driver. This lead to only the stubs
being added to the runpath of CUDA executables. The stubs are only meant
to be used at link time but not functional during runtime.

This change prepends the runpath with the correct CUDA driver.

Also removes the now obsolete `LL_CUDA_RUNTIME` flag.